refactor: input field const refactored to formelements
